+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Bugfixes
 
+* In the parser, fix `@links.@count` when applied over lists to return
+  the sum of backlinks for all connected rows in the list.
 * Fix null comparisons in queries not serialising properly in some cases.
   Also explicitly disable list IN list comparisons since its not supported.
   PR [#3037](https://github.com/realm/realm-core/pull/3037).